+++ apps/utr-backend/src/endpoints/entry/EntryService.ts
