+++ apps/utr-backend/src/endpoints/utils/AuthenticationUtils.ts
