+++ apps/utr-backend/src/endpoints/utils/EntryUtils.ts
